+++ changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-Last change: 06-MAR-2026 CEST
+Last change: 09-MAR-2026 CEST
 
 0.9.8
 -----
@@ -18,6 +18,7 @@ Last change: 06-MAR-2026 CEST
 - Added tool tips to G5M dialog
 - SPINNA GUI allows for user-defined max y-axis value in the NND plot
 - Fixed G5M 3D CI test
+- Localize GUI allows the user to automatically undrift localizations
 
 0.9.7
 -----